@@ -23,7 +23,7 @@ import static com.poorknight.house.commands.HouseCommand.AT_WORK_MODE;
 import static com.poorknight.house.commands.HouseCommand.DIM_LIGHTS;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Matchers.eq;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
